Update Project Layout; Create Tokenize Route
Update the project structure to pull the flask script out to the main folder. Create the Tokenizing route.
