+++ textTkAPI/__init__.py
@@ -1,0 +1,8 @@
+"""
+textTkAPI
+"""
+
+__version__ = '0.1.0'
+
+
+from .textProcessing import *
